requirements.txt
flask
pytube
requests
python-dotenv
faiss-cpu
langchain
langgraph
langchain_groq
sentence-transformers
